refactor: move function
- Jacobian creation is more seamless if integrated into the wrapper
- Avoids any additional complication regarding keeping track of function evaluations performed in the numerical jacobian
